requirements.txt
-iohttp==3.7.3
aiofiles==0.6.0
aiohttp==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.5.1
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
multidict==4.7.6
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
yarl==1.5.1
